pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "silo"
version = "0.9.0"
description = "A unified CLI for local AI models on Apple Silicon"
readme = "PROJECT.md"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "typer>=0.15",
    "rich>=13.0",
    "pydantic>=2.12",
    "pyyaml>=6.0",
    "huggingface-hub>=0.28",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sse-starlette>=2.0",
    "python-multipart>=0.0.20",
]

[project.optional-dependencies]
mlx = [
    "mlx>=0.22",
    "mlx-lm>=0.22",
]
audio = [
    "mlx>=0.22",
    "mlx-whisper>=0.4",
    "mlx-audio>=0.2",
]
llamacpp = [
    "llama-cpp-python>=0.3",
]
tui = [
    "textual>=0.50",
]
flows = [
    "prefect>=3.0",
]
discovery = [
    "zeroconf>=0.132",
]

[dependency-groups]
dev = [
    "pytest>=9.0",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.25",
    "httpx>=0.28",
    "ruff>=0.11",
    "mypy>=1.15",
]

[project.scripts]
silo = "silo.cli.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/silo"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
